requirements.txt
fastapi
uvicorn
kafka-python
pydantic
python-dotenv
apache-airflow
dbt-core
dbt-duckdb 
duckdb
spark
psycopg2
aiokafka